experiments/imagenet_classification/requirements.txt
# --------- basics --------- #
numpy==2.0.0
scipy==1.14.0
pandas==2.2.2
torch==2.3.1
torchvision==0.18.1
pytorch-ignite==0.4.13

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
omegaconf==2.3.0

# --------- loggers --------- #
tqdm==4.66.2
tensorboardX==2.6.2.2
tensorboard==2.16.2

# --------- compression --------- #
lrf @ git+https://github.com/pashtari/lrf.git

# --------- visualization --------- #
matplotlib==3.8.4
seaborn==0.13.2

# --------- misc --------- #
pyinstrument==4.6.2